Fix for full personal shields still taking charge and not being disabled when selected on the building.

+++ ED-Core/Source/Enhanced_Defence/PersonalShields/ThingDef_PersonalNanoShield.cs
@@ -12,7 +12,7 @@ namespace Enhanced_Defence.PersonalShields
         //public float energyLossPerDamage = 0.027f;
         public float minDrawSize = 1.2f;
         public float maxDrawSize = 1.55f;
-        public float maxEnergy = 100;
+        public int maxEnergy = 100;
         public bool isRotating = true;
         public String bubbleGraphicPath = "Other/ShieldBubble";
         public String soundAbsorb = null;
